requirements.txt
openai
langchain
faiss-cpu
tiktoken
python-dotenv
duckduckgo-search
streamlit-auth0-component
